Gets signed in user's canvas access token from their profile, after their account is linked.
Professor Wang requested for us to implement an Authentication UI that allows the user to sign in and provide our
Skill their access token. The Authentication UI is hosted by Amazon Cognito and the Alexa project makes a HTTP request to get user info from our Cognito User Pool. Because the text field for the custom attribute we created doesnot show on the sign-up form, we used zoneinfo (one of their built-in attributes) to hold the access token. Aftergetting the JSON response, the Skill extracts the access token from zoneinfo and creates the header that will be used for all Canvas API calls.